UPSTREAM: drivers/i2c/tpm: Work around missing firmware_version in Ti50 < 0.0.15
Ti50 firmware versions below 0.0.15 don't support the firmware_version
register and trying to access it causes I2C errors. Some nissa boards
are still using Ti50 0.0.12, so add a workaround Kconfig to skip reading
the firmware version and select it for nissa. The firmware version is
only read to print it to the console, so it's fine to skip this. This
workaround will be removed once all ODM stocks are updated to 0.0.15 or
higher.

A similar workaround Kconfig was added in CB:63011 then removed in
But we actually still need this workaround until all Ti50 stocks are
upgraded to 0.0.15 or higher.

BUG=b:224650720
TEST=Boot to OS on nereid with Ti50 0.0.14

(cherry picked from commit d88233ecd34605031c41cdf3aeb50bbd82a442d9)

CB:63158 which added support for separate handling of Cr50 and Ti50.
Signed-off-by: Reka Norman <rekanorman@chromium.org>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/64202
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: Ia30d44ac231c42eba3ffb1cb1e6d83bb6593f926
GitOrigin-RevId: d88233ecd34605031c41cdf3aeb50bbd82a442d9
Change-Id: I6c2f7a33feb99347aedaa191d7e66401285fe5d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3650646
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Jonathon Murphy <jpmurphy@google.com>