UPSTREAM: commonlib: Add timestamp IDs for Chrome OS hypervisor
Chrome OS is experimenting with a hypervisor layer that boots after
firmware, but before the OS. From the OS' perspective, it can be
considered an extension of firmware, and hence it makes sense to emit
timestamp to track hypervisor boot latency. This change adds
timestamp IDs in the 1200-1300 range for this purpose.

BUG=b:217638034
BRANCH=none
TEST=Manual: cbmem -a TS_CRHV_BOOT to add a timestamp, cbmem -t to
verify that it got added to the timestamp table.

Signed-off-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/64226
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: If70447eea2c2edf42b43e0198b827c1348b935ea
GitOrigin-RevId: c8c6185d8eda1a58a1c7084812e93fb3d2c8f254
Change-Id: Ifa10aafa6e22c84668239338d41790b91ab1dc34
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3644035
Reviewed-by: Peter Marheine <pmarheine@chromium.org>
Commit-Queue: Peter Marheine <pmarheine@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>