UPSTREAM: arch/x86/car.ld: Add a Kconfig param to flag AGESA brokenness
AGESA has a lot of code in the .data section (initialized data). However
there is no such section in CAR stages as the code runs in XIP mode and
CAR is too small to contain the data section. When the linker can not
match code to a section it will just append it, which is why AGESA
worked at all.

Follow-up patches will attempt to fix AGESA and set Kconfig parameter to
'n'. After all AGESA sources have been fixed, this can be removed.

Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/64387
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-by: Kyösti Mälkki <kyosti.malkki@gmail.com>
Original-Change-Id: I311ee17e3c0bd283692194fcee63af4449583d74
GitOrigin-RevId: cd259cb08a99812082f8bb2f7ecd23c031bb4e2d
Change-Id: Ic08cb85d8d8182d1ac47f5acd8cc13ae00a5e814
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3657389
Reviewed-by: Rob Barnes <robbarnes@google.com>
Commit-Queue: Rob Barnes <robbarnes@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Reka Norman <rekanorman@chromium.org>