UPSTREAM: vendorcode/amd/agesa/f14: Fix improper use of .data
AGESA has a lot of code in the .data section which is for initialized
data, that in fact should be .rodata. This adds the 'CONST' keyword
everywhere it is needed.

.data section.

(cherry picked from commit 704ccafb392212cffc3810197da5cb2b619f1ba6)

Original-TEST: See in the .elf file (e.g. using readelf) that there is nothing in
Original-Change-Id: I657d09f05070f5a88a4a162872c961db869a8df3
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/64399
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-by: Stefan Reinauer <stefan.reinauer@coreboot.org>
GitOrigin-RevId: 704ccafb392212cffc3810197da5cb2b619f1ba6
Change-Id: I2e182e176d630ba7d74561be4628e83744745c0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3675258
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>