UPSTREAM: arch/x86: Make sure bootblock gets buildtested
Now that the bootblock isn't added to cbfs anymore, on some targets it's
only conditionally build. One example would be Intel APL where it only
gets build when stitched into an IFWI. This is always done when
compiling for real targets but not by the CI builder. This adds a dummy
target to make sure the bootblock always gets buildtested.

Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/64427
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Nico Huber <nico.h@gmx.de>
Reviewed-by: Werner Zeh <werner.zeh@siemens.com>
Original-Change-Id: I60601e01a2c370b5c21493b71d51f495bb42f41d
GitOrigin-RevId: db9873b69c52cc1eac5bfef72228f071e303a78f
Change-Id: I23c5e0073318be9473f17db1e88fc35436fb66e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3648642
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Reka Norman <rekanorman@chromium.org>
Reviewed-by: Reka Norman <rekanorman@chromium.org>