UPSTREAM: Documentation: Move intelp2m from description.md to Documention
The description.md file for the intelp2m utility wasn't the description
that was needed - just a subject, and what language it was written in.
It was instead a set of more full documentation, so move it into the
Documentation directory and create a new description file.

(cherry picked from commit 81e0d689c07818d6a6554240947e5ab16bf3afc4)

Original-Change-Id: Ia180ae41f91f8b8eb408351a9e44e899edc031d3
Original-Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/64578
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Felix Singer <felixsinger@posteo.net>
GitOrigin-RevId: 81e0d689c07818d6a6554240947e5ab16bf3afc4
Change-Id: I1d725e790f4872bfa5443deba2579dd935f9cacc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3676329
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>