UPSTREAM: intel/common/block: move RAPL disabling to common code
This patch brings the feature of disabling RAPL to common code. It
replaces the current solution for APL and EHL.
For special case if RAPL disabling is only working via changes in MCHBAR
a new config switch was introduced.

relevant bits in MSR 0x610 are the same as before the
patch.

(cherry picked from commit d2d90215435f05329bfc03d9351ee6f8f5351014)

Original-Test: Boot mc_apl4/5 with this patch and ensure that the
Original-Change-Id: I2098ddcd2f19e3ebd87ef00c544e1427674f5e84
Original-Signed-off-by: Uwe Poeche <uwe.poeche@siemens.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/64596
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-by: Lean Sheng Tan <sheng.tan@9elements.com>
GitOrigin-RevId: d2d90215435f05329bfc03d9351ee6f8f5351014
Change-Id: I30214731f3dc1138d1b7bc1b10689747038c54e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3689528
Commit-Queue: Nick Vaccaro <nvaccaro@google.com>
Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Nick Vaccaro <nvaccaro@google.com>