UPSTREAM: soc/intel/apollolake: Compare patched FIT pointer with the pre-defined
Since the FIT pointer is patched at runtime there is no guarantee that
the pre-defined one will match the patched one. Add a check and print a
warning at runtime if both addresses (pre-defined and patched) do not
match as in this case an offline computed hash for the bootblock will
differ from the runtime one.

Signed-off-by: Werner Zeh <werner.zeh@siemens.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/64598
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Change-Id: Ib1b02ec43af183caa9f5b08b3c485879b423c40f
GitOrigin-RevId: de1459082b08cf17c5e0c82fde5430801eec46ff
Change-Id: If6129d735a83a063241e43bc319a61827d875768
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3670889
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Rob Barnes <robbarnes@google.com>
Commit-Queue: Rob Barnes <robbarnes@google.com>