UPSTREAM: mb/starlabs/labtop: Define CCD Port in Kconfig
Define the CCD (aka "Webcam") USB port in the devicetree as
it is used in multiple places. It is used in devtree to
disable it based on the CMOS setting "webcam", and in the
devicetree to configure the port tuning.

This also corrects the port that is disabled on CML, from
usb2_port[6] to usb2_port[3].

(cherry picked from commit 2eb2dcebc7b09a315b4f7c6bdc8f4abf4220de61)

Original-Signed-off-by: Sean Rhodes <sean@starlabs.systems>
Original-Change-Id: I16e368fc7965f978f2302633122ba63038603c1e
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/64704
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: 2eb2dcebc7b09a315b4f7c6bdc8f4abf4220de61
Change-Id: Ie7649489db77ba3d7af12324b071c6e83427b945
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3748007
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Reka Norman <rekanorman@chromium.org>
Reviewed-by: Reka Norman <rekanorman@chromium.org>