UPSTREAM: cbfs: Add CBFS_TYPE_INTEL_FIT and exclude it from CBFS verification
The Intel Firmware Interface Table (FIT) is a bit of an annoying outlier
among CBFS files because it gets manipulated by a separate utility
(ifittool) after cbfstool has already added it to the image. This will
break file hashes created for CBFS verification.

This is not actually a problem when booting, since coreboot never
actually loads the FIT from CBFS -- instead, it's only in the image for
use by platform-specific mechanisms that run before coreboot's
bootblock. But having an invalid file hash in the CBFS image is
confusing when you want to verify that the image is correctly built for
verification.

This patch adds a new CBFS file type "intel_fit" which is only used for
the intel_fit (and intel_fit_ts, if applicable) file containing the FIT.
cbfstool will avoid generating and verifying file hashes for this type,
like it already does for the "bootblock" and "cbfs header" types. (Note
that this means that any attempt to use the CBFS API to actually access
this file from coreboot will result in a verification error when CBFS
verification is enabled.)

(cherry picked from commit af20fd748b54366813d58fcbc3fb33189f32f116)

Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: I1c1bb6dab0c9ccc6e78529758a42ad3194cd130c
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/64736
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
GitOrigin-RevId: af20fd748b54366813d58fcbc3fb33189f32f116
Change-Id: I0896604968ac980026ce133cfd3e76ac452d8ed5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3686897
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>