OWNERS: Replace `*` with cros_ap and chromeos_peng_sies
This patch reduces owner approval rights to all members of the AP
firmware and SIE teams (since these are primary contributors and
maintainers of this repo, including firmware branches).

This is a retry of CL:3485996, this time with owner list links that
should (hopefully) work. Removing `*` owners is necessary so that Rubber
Stamper starts working again for cherry-picks.

BRANCH=none
BUG=none
TEST=none

Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: I4b77cfe3795ab4c7f8915f22b937b2a905669c9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3687245
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>