UPSTREAM: external/LinuxBoot: Deprecate GOPATH in u-root
This is a breaking change for now when using latest u-root main, which
is the default behavior in LinuxBoot.
u-root switched to golang modules and therefore `go get` is not the
standard behavior anymore. The workaround for this is to pull the
repository and build directly in the directory for now. Another apporach
would be to use `go install $pkg@latest` to install the binary at that
particular version into the golang binary path.

Currently missing is a control structure to enable the build process for
legacy versions <v0.8.0.

(cherry picked from commit b35c1f45a91774f77a767d4aa1c21dea6f29f74c)

Original-Signed-off-by: Patrik Tesarik <patrik.tesarik@9elements.com>
Original-Change-Id: Ifa03504da6fa321ffc6d2506b27ebd2e3ed9961b
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/65090
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Christian Walter <christian.walter@9elements.com>
GitOrigin-RevId: b35c1f45a91774f77a767d4aa1c21dea6f29f74c
Change-Id: I02c7b1362ba130aa2184dbd127c28462d06fbb08
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3707192
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>