UPSTREAM: mb/starlabs/lite/glkr: Correct the daughterboard USB 3.0 port number
The daughterboard USB 3.0 was set to port 3, which is incorrect. This
patch corrects that to port 4.

This fixes an issue where USB 3.0 devices are not detected when plugged
in to this port.

(cherry picked from commit b02c90d146f4929acdee0f28e4c0ddd545ffdb04)

Original-Signed-off-by: Sean Rhodes <sean@starlabs.systems>
Original-Change-Id: I50f86dee1b512d0dd20d07e3ee17ebfa5e537bc9
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/65186
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
GitOrigin-RevId: b02c90d146f4929acdee0f28e4c0ddd545ffdb04
Change-Id: Ia80a19030abf8956c70f52baf3f57595e10ed3c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3716531
Commit-Queue: Rob Barnes <robbarnes@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Rob Barnes <robbarnes@google.com>