UPSTREAM: mb/clevo/l140mu: work around PECI staying high when idle, blocking s0ix
According to Intel doc# 575683 the PECI bus should be low when idle and
is pulled up by clients with strong drive. However, for unknown reasons
the bus stays high on this board, blocking s0ix entry.

The PECI reference schematic in the ASPEED AST2400 BMC datasheet
(actually not related to this board) says that a pull-down is *required*
for the idle state.

This might be just a requirement of this BMC, since this is nowhere
documented in Intel datasheets, schematics or elsewhere. However,
configuring a weak pull-down (20 k) on the PECI pad indeed solves this
problem for now.

(cherry picked from commit 3fa42ac553e0e7583a0e8db2e8c34fbd09a07225)

Original-Change-Id: I85193000af67cd2c0465bdbb58cdd51b68fd5b4f
Original-Signed-off-by: Michael Niewöhner <foss@mniewoehner.de>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/68794
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: 3fa42ac553e0e7583a0e8db2e8c34fbd09a07225
Change-Id: Id12f225b01fba2361e47c56b604f23b7e230e5cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4000736
Reviewed-by: Kangheui Won <khwon@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>