UPSTREAM: mb/clevo/l140cu: work around PECI staying high when idle, blocking s0ix
According to Intel doc# 575683 the PECI bus should be low when idle and
is pulled up by clients with strong drive. However, for unknown reasons
the bus stays high on this board, blocking s0ix entry.

The PECI reference schematic in the ASPEED AST2400 BMC datasheet
(actually not related to this board) says that a pull-down is *required*
for the idle state.

This might be just a requirement of this BMC, since this is nowhere
documented in Intel datasheets, schematics or elsewhere. However,
configuring a weak pull-down (20 k) on the PECI pad indeed solves this
problem for now.

(cherry picked from commit 8998ab8b020d6711f56634880b26b2b908d1340c)

Original-Change-Id: Ib5a6b0ad3553c2cf795037d6a1982102bcb04644
Original-Signed-off-by: Michael Niewöhner <foss@mniewoehner.de>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/68793
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: 8998ab8b020d6711f56634880b26b2b908d1340c
Change-Id: I28e06b222494882ce3d05d26e31f974f135c5b90
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4000735
Commit-Queue: Patrick Georgi <pgeorgi@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Kangheui Won <khwon@chromium.org>