UPSTREAM: mb/google/brya/acpi: Add support for D Notify event from the Chrome EC
The agah EC code includes a driver to keep track of the current D Notify
level that the GPU should be at. When it changes, it will send a host
event to the ACPI FW, which will then pass that Notify on to the kernel
driver. This patch adds support for that feature, which is described in
the Nvidia Software Design Guide.

BUG=b:229405562
TEST=add Printf() calls to the ACPI, and work through the various
scenarios on the EC that will cause D Notify levels to change; this
will cause the Printfs() to show up in the kernel log.

(cherry picked from commit 8702450e51114d6e366005ac78331661bc5ad93b)

Original-Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Change-Id: I5cd8bd7d177ea10a165613ed0726a6d6fd86c226
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/65486
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Robert Zieba <robertzieba@google.com>
Original-Reviewed-by: Paul Menzel <paulepanter@mailbox.org>
Original-Reviewed-by: Eric Lai <eric_lai@quanta.corp-partner.google.com>
GitOrigin-RevId: 8702450e51114d6e366005ac78331661bc5ad93b
Change-Id: I1852fb676f5794f24a243f2fbbf570697b0cb894
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3779967
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>