UPSTREAM: include: Add SPDX-License-Identifiers to files missing them
This adds SPDX-License-Identifiers to all of the files in src/include
that are missing them or have unrecognized identifiers.

Files that were written specifically for coreboot and don't have license
information are licensed GPL-2.0-only, which is the license for the
overall coreboot project.

Files that were sourced from Linux are similarly GPL-2.0-only.

The cpu/power files were committed with source that was licensed as
GPL-2.0-or-later, so presumably that's the license for that entire
commit.

The final file, vbe.h gives a pointer to the BSD-2-Clause license
at opensource.org.

(cherry picked from commit 239b5df2687e30636688354c852cdca1bafbc80b)

Original-Change-Id: I3f8fd7848ce11c1a0060e05903fb17a7583b4725
Original-Signed-off-by: Martin Roth <martin.roth@amd.corp-partner.google.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/66284
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Elyes Haouas <ehaouas@noos.fr>
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-by: Felix Singer <felixsinger@posteo.net>
GitOrigin-RevId: 239b5df2687e30636688354c852cdca1bafbc80b
Change-Id: I3086413e86f10a777c54d4cb550b83c46bee0602
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3804032
Tested-by: Nick Vaccaro <nvaccaro@google.com>
Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Commit-Queue: Nick Vaccaro <nvaccaro@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>