UPSTREAM: payloads/edk2: Move the restoration of the logo
Logo.bmp is overwritten with a custom one from coreboot. This needs to
be restored before the branch is updated otherwise git will report that
the repository is dirty.

Move this to the update recipe so that will always be done for any
recipe that needs to update the branch.

(cherry picked from commit b02cc14367890661ca047ffe8c5d86a3e6bac95d)

Original-Signed-off-by: Sean Rhodes <sean@starlabs.systems>
Original-Change-Id: I85bf753a47d9e70d6555dec9a539e8ed7395bead
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/66355
Original-Reviewed-by: Lean Sheng Tan <sheng.tan@9elements.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: b02cc14367890661ca047ffe8c5d86a3e6bac95d
Change-Id: Ice6aca7ad950718c70441d2d0946d9d58a884926
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3831915
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Rob Barnes <robbarnes@google.com>
Reviewed-by: Rob Barnes <robbarnes@google.com>