UPSTREAM: util/lint & LICENSES: Add PDDC as a "license" for coreboot
The Creative Commons Public Domain Dedication and Certification is not
a license in the common sense in that it's stating that the associated
file is already in the public domain (having no copyright), and is not
actually putting it in the public domain like the CC-C0 license does.

The use for this in coreboot is for unlicensable files - either blank
files or files with no creative content.  This allows these files to
have the SPDX identifier to identify them as having no known copyright
for open source license compliance.

If CC-PDM-1.0 is ever included in the list of SPDX licenses, that would
probably fit better, but because the public domain mark isn't actually
a license, and because "public domain" isn't well defined, CC-PDM was
rejected as a SPDX identifier.

For further information:
https://web.archive.org/web/20201018194411/https://github.com/spdx/license-list-XML/issues/988

(cherry picked from commit f67a1aa76af637b8fb9583dcdffe0a2ef4e31940)

Original-Change-Id: Ibb300ecd066cde2a016195c2beca76a460c588e3
Original-Signed-off-by: Martin Roth <martin.roth@amd.corp-partner.google.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/66496
Original-Reviewed-by: Elyes Haouas <ehaouas@noos.fr>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Michael Niewöhner <foss@mniewoehner.de>
GitOrigin-RevId: f67a1aa76af637b8fb9583dcdffe0a2ef4e31940
Change-Id: I517f74b7b6d5d17321bb326aa32d133c190f1aab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3826104
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>