UPSTREAM: pciexp_device: Join pciexp_find_(next_)extended_cap() APIs
Move the `offset` parameter into pciexp_find_extended_cap(). If it's
called with `0`, we start a new search. If it's an existing offset,
we continue the search.

This makes it easier to search for multiple occurences of a capa-
bility in a single loop.

(cherry picked from commit 5ffc2c8a3f38fbb7be2faadf207590acd3999205)

Original-Change-Id: I80115372a82523b90460d97f0fd0fa565c3f56cb
Original-Signed-off-by: Nico Huber <nico.h@gmx.de>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/66453
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
GitOrigin-RevId: 5ffc2c8a3f38fbb7be2faadf207590acd3999205
Change-Id: I6efde2811dac0d515e205bcf63f77313603c37ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3837870
Commit-Queue: Kangheui Won <khwon@chromium.org>
Reviewed-by: Kangheui Won <khwon@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>