UPSTREAM: pciexp_device: Properly search for Intel's 0xcafe capability
We have this quirk in our tree since the introduction of L1-substate
support[1]. The way we searched for this capability was rather crude:
We simply assumed that it would show up in the first data word of
another capability.

As it turned out that it is actually a proper vendor-specific capa-
bility that we are looking for, we can drop some of the mystic code.
This was confirmed to work on the device that was originally used
during development, Google/Samus.

[1] commit 31c6e632cf (PCIe: Add L1 Sub-State support.)

(cherry picked from commit bba97354b05cc8770d6d913bfe5777b46450bbb8)

Original-Change-Id: I886fb96e9a92387bc0e2a7feb746f7842cee5476
Original-Signed-off-by: Nico Huber <nico.h@gmx.de>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/66455
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
GitOrigin-RevId: bba97354b05cc8770d6d913bfe5777b46450bbb8
Change-Id: I39cf11e1cd2118497bdc98b94d281890b87a65f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3837992
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Kangheui Won <khwon@chromium.org>
Reviewed-by: Kangheui Won <khwon@chromium.org>