UPSTREAM: vc/amd/fsp/cezanne,mendocino: add FSP CCX CPPC HOB GUID and struct
To generate a complete _CPC ACPI object, coreboot needs the minimal and
nominal core speed values which are specific to the CPU and not only the
CPU family. Since this is done by an undocumented mechanism, FSP has to
do this and puts the information we need into a HOB. This adds the HOB
GUID and the structure of the HOB data.

(cherry picked from commit b65845cb2b0009d1ed1e9b8550cbcd4c1d093c63)

Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Signed-off-by: Matt DeVillier <matt.devillier@amd.corp-partner.google.com>
Original-Change-Id: Ibf338c32de367a3fd57695873da1625338fa196d
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/66549
Original-Reviewed-by: Raul Rangel <rrangel@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: b65845cb2b0009d1ed1e9b8550cbcd4c1d093c63
Change-Id: Ia158b28903c1895d997c4a31d1086dac9d9aa871
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3831197
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Rob Barnes <robbarnes@google.com>
Commit-Queue: Rob Barnes <robbarnes@google.com>